requirements.txt
#GetOldTweets3
gensim
jupyter
lightgbm
matplotlib
nltk
#nb_conda
numpy
pandas
regex
scikit-learn
seaborn
tensorflow-gpu
tensorflow-hub
wordcloud
#nb_conda_kernels
#ipykernel
#notebook==5.7.4